ops/site-audit/requirements.txt
PyYAML==6.*
requests==2.*
beautifulsoup4==4.*
jsonschema==4.*
